requirements.txt

numpy
pandas
streamlit
pinecone-client
python-dotenv
scikit-learn
langchain-openai
langchain_pinecone
langchain_community
